fix(completions): generate v3 of alloy and titanium for completions
