+++ src/extension.ts
@@ -1,19 +1,18 @@
-import * as path from 'path';
 import * as vscode from 'vscode';
 import appc from './appc';
 import project from './project';
 
-import { Commands, handleInteractionError, InteractionChoice, InteractionError } from './commands';
+import { Commands, handleInteractionError, InteractionError } from './commands';
 import { GlobalState } from './constants';
 import { ExtensionContainer } from './container';
 
 import { Config, Configuration, configuration } from './configuration';
 
-import { completion, environment, updates } from 'titanium-editor-commons';
+import { environment } from 'titanium-editor-commons';
 
 import { registerTaskProviders } from './tasks/tasksHelper';
 import { registerDebugProvider } from './debugger/titaniumDebugHelper';
-import { registerProviders } from './providers';
+import { generateCompletions, registerProviders } from './providers';
 import { registerCommands } from './commands/index';
 import { registerViews } from './explorer';
 
@@ -38,97 +37,6 @@ export async function activate (context: vscode.ExtensionContext): Promise<void>
 
 export function deactivate (): void {
 	project.dispose();
-}
-
-/**
- * Generate Alloy and Titanium SDK Completion files
- *
- * @param {boolean} [force=false] generate the completions even if they exist
- */
-async function generateCompletions (force = false): Promise<void> {
-	if (!project.isValid()) {
-		return;
-	}
-	let sdkVersion: string|string[]|undefined;
-	try {
-		sdkVersion = project.sdk();
-		if (!sdkVersion) {
-			const error = new InteractionError('Errors found in tiapp.xml: no sdk-version found');
-			error.interactionChoices.push({
-				title: 'Open tiapp.xml',
-				run: async () => {
-					const file = path.join(vscode.workspace.rootPath!, 'tiapp.xml');
-					const document = await vscode.workspace.openTextDocument(file);
-					await vscode.window.showTextDocument(document);
-				}
-			});
-			throw error;
-		} else if (sdkVersion.length > 1) {
-			const error = new InteractionError('Errors found in tiapp.xml: multiple sdk-version tags found.');
-			error.interactionChoices.push({
-				title: 'Open tiapp.xml',
-				run: async () => {
-					const file = path.join(vscode.workspace.rootPath!, 'tiapp.xml');
-					const document = await vscode.workspace.openTextDocument(file);
-					await vscode.window.showTextDocument(document);
-				}
-			});
-			throw error;
-		} else {
-			sdkVersion = sdkVersion[0];
-		}
-
-	} catch (error) {
-		if (error instanceof InteractionError) {
-			await handleInteractionError(error);
-		}
-		return;
-	}
-	try {
-		const sdkInfo = appc.sdkInfo(sdkVersion);
-		if (!sdkInfo) {
-			// TODO
-			return;
-		}
-		const sdkPath = sdkInfo.path;
-		// Generate the completions
-		const [ alloy, sdk ] = await Promise.all([
-			completion.generateAlloyCompletions(force, completion.CompletionsFormat.v2),
-			completion.generateSDKCompletions(force, sdkVersion, sdkPath, completion.CompletionsFormat.v3)
-		]);
-		if (sdk || alloy) {
-			let message = 'Autocomplete suggestions generated for';
-			if (sdk) {
-				message = `${message} Titanium ${sdk}`;
-			}
-			if (alloy) {
-				message = `${message} Alloy ${alloy}`;
-			}
-			vscode.window.showInformationMessage(message);
-		}
-	} catch (error) {
-		const actions: InteractionChoice[] = [];
-		if (error.code === 'ESDKNOTINSTALLED') {
-			actions.push({
-				title: 'Install',
-				run: () => {
-					vscode.window.withProgress({ location: vscode.ProgressLocation.Notification, title: 'Titanium SDK Installation', cancellable: false }, async () => {
-						try {
-							await updates.titanium.sdk.installUpdate(sdkVersion as string);
-							await appc.getInfo();
-							await generateCompletions(force);
-						} catch (err) {
-							return Promise.reject(err);
-						}
-					});
-				}
-			});
-		}
-		const install = await vscode.window.showErrorMessage(`Error generating autocomplete suggestions. ${error.message}`, ...actions);
-		if (install) {
-			await install.run();
-		}
-	}
 }
 
 /**
